% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoeocc.R
\name{horseshoeocc}
\alias{horseshoeocc}
\title{horseshoeocc: Bayesian occupancy modeling with the regularized horseshoe prior in R}
\description{
\if{html}{\figure{logo.png}{options: align='right' width=120px}}

This package allows for the fitting and summarizing of the
occupancy models using the regularized horseshoe prior. We use
the notation used by \href{https://doi.org/10.1214/17-EJS1337SI}{Piironen and Vehtari (2017)}.
This package allows the user to specify a prior on the guess for the number
of relevant coefficients (p0).
}
\details{
Models are fit using JAGS and therefore require a local JAGS installation.
}
\seealso{
Useful links:
\itemize{
\item \href{https://doi.org/10.1890/0012-9658(2002)083\link{2248:ESORWD}2.0.CO;2}{MacKenzie et al. (2002) occupancy model paper}
\item \href{https://doi.org/10.1214/17-EJS1337SI}{Piironen and Vehtari (2017) regularized horseshoe prior paper}
}
}
\author{
\strong{Maintainer:} Meaghan Winder \email{mwinder@middlebury.edu} \cr \cr
Authors:
\itemize{
\item Meaghan Winder
\item Dr. Andrew Hoegh
}
}
