% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{koala_data100}
\alias{koala_data100}
\title{Koala occupancy data}
\format{
A list with 2 elements:
\describe{
  \item{site_df}{A data frame with site information, containing 123 rows and
  101 columns. Columns include the site identifier (\code{site}) and the
  site-level covariates (\code{LTC}, \code{SRD}, and \code{x3}:\code{x100}).}
  \item{samp_df}{A data frame with sample information, containing 911 rows
  and 6 columns. Columns include site identifier (\code{site}), sample
  identifier (\code{samp}), the sample-level covariate (\code{weather}), in addition
  to the two site-level covariates (\code{LTC} and \code{SRD}). The data frame
  also includes detection/non-detection data (\code{y}, where 1 indicates detection and
  0 indicates non-detection).}

}
}
\usage{
koala_data100
}
\description{
We use koala occupancy data from Mornington Peninsula, Victoria,
Australia collected by
\href{https://doi.org/10.1007/s10980-023-01620-2}{Whisson et al. (2023)}.
Whisson et al. (2023) investigate koala occupancy of 123 sites using acoustic surveys,
and found that tree cover (LTC) and road density within 1 km of the
recording device (SRD) were associated with changes in site occupancy rates;
they estimated that site occupancy increased as tree cover increased,
and declined as road density increased. Additionally, Whisson et al. (2023) found
that detection probability decreased with increased inclement weather. We use the
data from Whisson et al. (2023) and add spurious covariates to create a synthetic data
examples with 100 potential occupancy covariates. The synethic data includes both
tree cover and road density; the remaining spurious covariates are generated
from a standard normal distribution.
}
\details{
This data set includes the site-level covariates found to be associated with changes
in site occupancy probabilities by Whisson et al. (\code{LTC} and \code{SRD}),
along with 98 simulated spurious site-level covariates (\code{x3}:\code{x100})
to demonstrate shrinkage under the regularized horseshoe prior. Additionally,
the data include a single sample-level covariate (\code{weather}) for the detection
probability of each sample.
}
\examples{
data(koala_data100)

}
\keyword{datasets}
