% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.horseshoeocc.R
\name{plot.horseshoeocc}
\alias{plot.horseshoeocc}
\title{Trace plots for occupancy model parameters fit using the regularized horseshoe prior}
\usage{
\method{plot}{horseshoeocc}(fit, which = c("beta"), ...)
}
\arguments{
\item{fit}{an object of class \code{horseshoeocc}, typically a result of
a model fit using \code{\link{horseshoeocc}}.}

\item{which}{the parameters for which to generate trace plots. The value can
be a specific parameter (i.e. \code{"beta[1]"}), or you can plot the trace plots
for all alpha, beta, lambda, or kappa coefficients using \code{"alpha"},
\code{"beta"}, \code{"lambda"}, or \code{"kappa"}, respectively; defaults to \code{"beta"}.}

\item{...}{Other arguments.}
}
\value{
Returns a ggplot object that is a trace plot of the chosen model parameter(s).
}
\description{
Plot method for objects of class \code{horseshoeocc}.
}
\details{
You may only select one group of parameters (i.e. \code{"beta"},
\code{"alpha"}, \code{"lambda"}, or \code{"kappa"}) to plot at a time.
}
\examples{
# continuing the horseshoeocc example:
## trace plots for the alpha coefficients
plot(fit, "alpha")

## trace plots for the non-zero beta coefficients
plot(fit, c("beta0", "beta[1]", "beta[2]"))

## trace plot for meff (either works)
plot(fit, "mefft")
plot(fit, "meff")

}
