% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.horseshoeocc_derivedsum.R
\name{plot.horseshoeocc_derivedsum}
\alias{plot.horseshoeocc_derivedsum}
\title{Title}
\usage{
\method{plot}{horseshoeocc_derivedsum}(x, which = c("psi"), hdi = FALSE, equal = NULL, ...)
}
\arguments{
\item{x}{tmp}

\item{which}{tmp}

\item{hdi}{tmp}

\item{equal}{tmp}

\item{...}{tmp}
}
\value{
plot
}
\description{
Title
}
