% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.horseshoeocc_sum.R
\name{plot.summary.horseshoeocc}
\alias{plot.summary.horseshoeocc}
\title{Credible interval plots for occupancy model parameters fit using the regularized horseshoe prior}
\usage{
\method{plot}{summary.horseshoeocc}(x, which = c("beta"), hdi = FALSE, equal = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{summary.horseshoeocc}, typically the
result of a call to \code{\link{summary.horseshoeocc}}.}

\item{which}{the parameters for which to generate trace plots. The value can
be a specific parameter (i.e. \code{"beta[1]"}), or you can plot the trace plots
for all of the beta coefficients using \code{"beta"}; defaults to \code{"beta"}.}

\item{hdi}{logical; defaults to \code{FALSE}. If \code{TRUE}, the highest
(posterior) density intervals (HDI) are plotted.}

\item{equal}{optional value to compare to each credibility interval.}

\item{...}{Other arguments.}
}
\value{
Returns a ggplot.
}
\description{
Credible interval plots for occupancy model parameters fit using the regularized horseshoe prior
}
