% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.summary.horseshoeocc_derived.R
\name{plot.summary.horseshoeocc_derived}
\alias{plot.summary.horseshoeocc_derived}
\title{Credible interval plots for occupancy model derived parameters fit using the regularized horseshoe prior}
\usage{
\method{plot}{summary.horseshoeocc_derived}(x, which = c("psi"), median = TRUE, hdi = TRUE, equal = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{summary.horseshoeocc_derived}, typically the
result of a call to \code{\link{summary.horseshoeocc_derived}}.}

\item{which}{the parameters for which to generate trace plots. The value can
be either \code{"psi"} or \code{"p"}; defaults to \code{"psi"}.}

\item{median}{logical; defaults to \code{TRUE}. If \code{FALSE}, the point
corresponds to the mean rather than the median.}

\item{hdi}{logical; defaults to \code{TRUE}. If \code{FALSE}, the quantile based
credibility intervals are used.}

\item{equal}{optional value to compare to each credibility interval.}

\item{...}{Other arguments.}
}
\value{
A ggplot (if \code{"psi"}) or list of ggplot objects (if \code{"p"}).
The ggplot objects are credible interval plots of the estimated occupancy or
detection probabilities.
}
\description{
Plot method for objects of class \code{summary.horseshoeocc_derived}.
}
\examples{
# continuing the summary.horseshoe_derived example:
## plot of credibility intervals for psi
plot(derived_summary, "psi")

## list of plots of credibility intervals for p
p_plots <- plot(derived_summary, "p")
p_plots[[1]]

}
