% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data100}
\alias{sim_data100}
\title{Simulated occupancy data}
\format{
A list with 6 elements:
\describe{
  \item{site_df}{A data frame with site information, containing 100 rows and
  101 columns. Columns include the site identifier (\code{site}) and the
  site-level covariates (\code{x1}:\code{x100}).}
  \item{samp_df}{A data frame with sample information, containing 1000 rows
  and 4 columns. Columns include site identifier (\code{site}), sample
  identifier (\code{samp}), the sample-level covariate (\code{w1}),
  and detection/non-detection data (\code{y}, where 1 indicates detection and
  0 indicates non-detection).}
  \item{z}{A vector containing the true occupancy state for each site (1 for
  occupied, 0 for unoccupied).}
  \item{x}{A data frame with 100 rows and columns \code{x1}:\code{x100},
  representing the site-level covariates for each of the 100 sites.}
  \item{w}{A data frame with 1000 rows and one column (\code{w1}), representing
  the sample-level covariate for each sample from each site.}
  \item{truth}{A list containing true parameter values for the models,
  assuming a logit link:
  \describe{
     \item{\code{beta0}}{Intercept for the occupancy probability model.}
     \item{\code{beta}}{Coefficients for the site-level covariates.}
     \item{\code{alpha0}}{Intercept for the detection probability model.}
     \item{\code{alpha}}{Coefficients for the sample-level covariate.}
  }}

}
}
\usage{
sim_data100
}
\description{
A simulated occupancy data set where there are 100 sites with 10 samples
collected per site. The data set is designed to illustrate the use of the
regularized horseshoe prior with a mixture prior on the prior guess for the
effective number of coefficients (\code{p0}).

This data set includes three true site-level covariates (\code{x1}:\code{x3})
that influence occupancy probability, along with 97 spurious site-level covariates
(\code{x4}:\code{x100}) to demonstrate shrinkage under the regularized horseshoe prior.
Additionally, the data include a single sample-level covariate (\code{w1}) for
the detection probability of each sample.
}
\examples{
data(sim_data100)

}
\keyword{datasets}
