% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoeocc.R
\name{horseshoeocc}
\alias{horseshoeocc}
\title{Fit occupancy model using the regularized horseshoe prior}
\usage{
horseshoeocc(
  data = list(site_df = NULL,
              samp_df = NULL),
  occ_model = ~.,
  det_model = ~1,
  priors = list(
    beta0 = list(mu0 = 0, sigma0 = 2),
    alpha0 = list(mu0 = 0, sigma0 = 2),
    alpha = list(mu0 = 0, sigma0 = 2),
    c = list(df_slab = 4, scale_slab = 2)
  ),
  mixture_mean1 = NULL,
  mixture_var1 = 1,
  mixture_weight1 = 0.5,
  mixture_mean2 = NULL,
  mixture_var2 = 25,
  mixture_weight2 = 0.5,
  pseudovar = 4,
  nchain = 3,
  niter = 1e+05,
  nburnin = niter/2,
  thin = 5,
  seed = NULL
)
}
\arguments{
\item{data}{an object of class \code{list} containing the following elements:
\itemize{
  \item \code{site_df} object of class \code{data.frame} containing site-specific
 covariates; \code{site_df} should have exactly one row for each site and have
 a column named \code{site}.
  \item \code{samp_df} object of class \code{data.frame} containing sample-specific
 covariates; \code{samp_df} should have exactly one row for each sample and
 have columns named \code{site} and \code{sample}.
}}

\item{occ_model}{formula describing the site-level model. The default \code{~.}
includes all columns in \code{site_df} except \code{site}.}

\item{det_model}{formula describing the sample-level model.}

\item{priors}{an object of class \code{list} containing the following elements:
 \itemize{
    \item \code{beta0} \verb{ }object of class \code{list} containing the
 following elements:
    \itemize{
      \item \code{mu0} \verb{ }prior mean for the site-level intercept
      \item \code{sigma0} \verb{ }prior standard deviation for the site-level intercept
    }
    \item \code{alpha0} \verb{ }object of class \code{list} containing the
 following elements:
    \itemize{
      \item \code{mu0} \verb{ }prior mean for sample-level intercept
      \item \code{sigma0} \verb{ }prior standard deviation matrix for sample-level intercept
    }
   \item \code{alpha} \verb{ }object of class \code{list} containing the
 following elements:
   \itemize{
     \item \code{mu0} \verb{ }prior mean for sample-level regression coefficients \cr
     \item \code{sigma0} \verb{ }prior standard deviation for sample-level regression
     coefficients
   }
   \item \code{c} \verb{ }object of class \code{list} containing the
 following elements:
   \itemize{
     \item \code{df_slab} \verb{ }prior degrees of freedom for the Student-t slab  \cr
     \item \code{scale_slab} \verb{ }prior scale of the Student-t slab
   }
}}

\item{mixture_mean1}{the prior mean of the Normal distribution truncated
between 1 and D-1 that serves as the first component of the mixture prior.}

\item{mixture_var1}{the prior variance of the Normal distribution truncated
between 1 and D-1 that serves as the first component of the mixture prior;
defaults to 1.}

\item{mixture_weight1}{the weight of the first component of the mixture prior;
defaults to 0.5.}

\item{mixture_mean2}{the prior mean of the Normal distribution truncated
between 1 and D-1 that serves as the second component of the mixture prior.}

\item{mixture_var2}{the prior variance of the Normal distribution truncated
between 1 and D-1 that serves as the second component of the mixture prior;
defaults to 25.}

\item{mixture_weight2}{the weight of the second component of the mixture prior;
defaults to 0.5.}

\item{pseudovar}{estimated value for the pseudo-variance; defaults to 4.}

\item{nchain}{number of Markov chains; defaults to 3.}

\item{niter}{number of iterations per chain (including burn-in); defaults to 100,000.}

\item{nburnin}{length of burn-in, i.e. number of iterations to discard at the
beginning. The default is \code{niter/2}, discarding the first half of
the simulations.}

\item{thin}{thinning rate. Must be a positive integer. Set \code{nthin} > 1 to
save memory and computation time if \code{niter} is large; defaults to 5.}

\item{seed}{an optional seed.}
}
\value{
An object of class \code{list} containing the elements:
\itemize{
  \item \code{mcmc} \verb{ }object of class \code{list} containing the posterior
  samples for each parameter. The number of elements corresponds to \code{nchain}.
  \item \code{data} \verb{ }object of class \code{list} containing:
    \itemize{
      \item \code{x} \verb{ }object of class \code{data.frame} containing the
      site-level covariates used in the model for occupancy probability.
      \item \code{w} \verb{ }object of class \code{data.frame} containing the
      sample-level covariates used in the model for detection probability.
      \item \code{site_df} \verb{ }inherited from input: \code{data$site_df}.
      \item \code{samp_df} \verb{ }inherited from input: \code{data$samp_df}.
    }
  \item \code{model} \verb{ }object of class \code{list} containing:
    \itemize{
      \item \code{det_model} \verb{ }formula describing the sample-level model.
      \item \code{occ_model} \verb{ }formula describing the site-level model.
    }
  \item \code{opts} \verb{ }object of class \code{list} containing the MCMC options:
  \code{niter}, \code{nchain}, \code{nburnin}, and \code{thin}.
}
}
\description{
This function fits a the single-season single species occupancy
model described by
\href{https://doi.org/10.1890/0012-9658(2002)083[2248:ESORWD]2.0.CO;2}{MacKenzie et al. (2002)}
using the regularized horseshoe prior described by
\href{https://doi.org/10.1214/17-EJS1337SI}{Piironen and Vehtari (2017)}.
Instead of specifying a value for the guess for the number of relevant coefficients (\eqn{p_0}),
the user should specify a mixture of truncated Normal distributions as a prior
distribution for \eqn{p_0}.
}
\examples{
# simulated dataset
data(sim_data100)

# fit the model to simulated data
fit <- horseshoeocc(data = sim_data100,
                    det_model = ~w1,
                    occ_model = ~.,
                    mixture_mean1 = 3,
                    mixture_var1 = 1,
                    mixture_mean2 = 3,
                    mixture_var2 = 25,
                    niter = 1000,
                    seed = 123)


}
